- place appropriate classes in corresponding their targets
